+++ CHANGELOG.md
@@ -1,6 +1,33 @@
 # Changelog
+All notable changes to Scriptable Object Menu will be documented in this file.
 
-## [1.0.3] - 2021-12-20
-Added ScriptableObject Menu as a Unity package.
-Added dialog path synchronisation with the Project Window.
-Added assembly filtering.
+The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
+## [2.0.0] - 2023-01-29
+
+### Added
+- Shortcut to create assets from selected types in the Project Window
+- Shortcut to clone assets from selected types in the Project Window
+- Ability to open newly created scripts with associated application
+- Ability to sort asset menu items alpha-numerically
+- Ability to group asset menu items by assembly
+- Editor only functions to the default template file
+- Configurable properties to the Project Settings window
+- Full support for user defined assemblies
+
+### Changed
+- Optimised type discovery via TypeCache (for v19.2+ users)
+- Refactored for full integration with the Project Window
+
+### Removed
+- Access modifiers from the default template file
+
+## [1.3.0] - 2021-12-20
+
+### Added
+- Unity package support
+- Assembly filtering
+- Dialog path synchronisation with the Project Window
+
+## [1.0.0] - 2019-12-24
+- Initial commit